Start splitting AppLogic into AppLogic and Window logic
  We'll need this for #5000, for ainulindale. This refactoring will be annoying
  enough as it is so we may as well do it as a first, separate PR.
